refactor(forms): extract FormFillFlow so every form viewer shares one behavior
